pyproject.toml
[project]
name = "ma-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "jupyter>=1.1.1",
    "langchain>=1.0.1",
    "langchain-core>=1.0.0",
    "langchain-huggingface>=1.0.0",
    "langchain-mcp-adapters>=0.1.11",
    "langchain-openai>=1.0.0",
    "langgraph>=1.0.1",
    "pyyaml>=6.0.3",
]
